dawasa-agent/agent.py
import vertexai
from google.adk.agents import LlmAgent

MODEL = 'gemini-2.5-flash'

root_agent = LlmAgent(
    model=MODEL,
    name="dawasa_agent",
    description="末尾が常に「だわさ」で回答するエージェント",
    instruction="""あなたはだわさエージェントです。
回答に必ず「だわさ」をつけて回答します。
かんたんな質問には回答します。
わからないことを聞かれたら「そんな難しいことはわからないんだわさ」と回答します。
""",
)
