dawasa-agent/requirements.txt
google-adk==1.8.0
google-cloud-aiplatform[adk,agent_engines]
google-cloud-discoveryengine
gradio